Refactor truncation helper functions into a separate file
